Adopt canonical lint config + Swift Testing in TailwindKit (#69)
Make the sources and tests clean under the package's own (canonical)
swift-format + SwiftLint config, both --strict, and convert the tests to the
house Swift Testing style.

- Shades: rename `Shade` cases `_50`…`_950` → `s50`…`s950` (swift-format's
  NoLeadingUnderscores forbids the leading underscore); call sites read
  `.bg(.blue, .s500)`. Updated all references, docs, README, CLAUDE.
- missing_docs / AllPublicDeclarationsHaveDocumentation: document every public
  static mirror and the Spacing/Size literal inits.
- file_length (225): split the two long files — bare instance utilities →
  `TailwindStyle+Bare.swift`, bare static mirrors → `TailwindStyle+StaticBare.swift`
  (parameterized methods stay in `+Utilities`/`+Static`). Filenames keep the
  `TailwindStyle+*` prefix so file_name passes.
- Tests → Swift Testing: `@Suite`/`@Test`/`#expect`, empty_string via `.isEmpty`,
  and avoid identical_operands by binding operands to locals. 15 tests pass.
